BWS-VPSC-V1 Code
This is the 1st generation code, I will be doing further research and development to improve the function of the code to control the pump in a more efficient and measured manner. In an attempt to give any potential user of my board or their own board the ability to control the power steering how they desire.